Fix incorrect return type

+++ src/support/types.h
@@ -125,7 +125,7 @@ typedef enum OMPTMVersion {
 
 static inline int min_int(int a, int b) { return a < b ? a : b; }
 static inline int max_int(int a, int b) { return a > b ? a : b; }
-static inline int max_int64(int64_t a, int64_t b) { return a > b ? a : b; }
+static inline int64_t max_int64(int64_t a, int64_t b) { return a > b ? a : b; }
 
 int  parse_bool(const char *str, bool *value);
 bool equivalent_bool(const char *str1, const char *str2);
